http: #25 rather not overwrite, too unsafe

+++ src/http_server.rs
@@ -407,8 +407,10 @@ pub async fn run(
 
                     trace!("file_path {}", sound_path.display());
 
-                    if let Err(err) = std::fs::write(&sound_path, upload_data) {
-                        return format_json_error(err);
+                    if !sound_path.exists() {
+                        if let Err(err) = std::fs::write(&sound_path, upload_data) {
+                            return format_json_error(err);
+                        }
                     }
 
                     let sound_config = config::SoundConfig {
